o1vm: reorganize with interpreters and "flavors"
