+++ o1vm/src/interpreters/keccak/tests.rs
@@ -1,124 +1,34 @@
 use crate::{
-    keccak::{
-        column::{
-            Absorbs::*,
-            Sponges::*,
-            Steps::{self, *},
-            N_ZKVM_KECCAK_COLS, N_ZKVM_KECCAK_REL_COLS, N_ZKVM_KECCAK_SEL_COLS,
-        },
+    interpreters::keccak::{
+        column::{Absorbs::*, Sponges::*, Steps::*},
         environment::KeccakEnv,
-        folding::KeccakConfig,
         interpreter::KeccakInterpreter,
-        trace::DecomposedKeccakTrace,
         Constraint::*,
         Error, KeccakColumn,
     },
     lookups::{FixedLookupTables, LookupTable, LookupTableIDs::*},
-    trace::{DecomposableTracer, Tracer},
-    BaseSponge, Fp,
+    Fp,
 };
-use ark_bn254::g1::Parameters;
+
 use ark_ff::{One, Zero};
-use folding::{
-    checker::{ExtendedProvider, Provider},
-    decomposable_folding::DecomposableFoldingScheme,
-    expressions::FoldingCompatibleExprInner,
-    FoldingCompatibleExpr, FoldingConfig, FoldingScheme,
-};
 use kimchi::{
-    circuits::polynomials::keccak::{constants::RATE_IN_BYTES, Keccak},
+    circuits::polynomials::keccak::Keccak,
     o1_utils::{self, FieldHelpers, Two},
 };
-use kimchi_msm::test::test_completeness_generic_no_lookups;
-use mina_poseidon::{constants::PlonkSpongeConstantsKimchi, sponge::DefaultFqSponge};
-use rand::{rngs::StdRng, Rng};
+use rand::Rng;
 use sha3::{Digest, Keccak256};
-use std::{
-    array,
-    collections::{BTreeMap, HashMap},
-};
-use strum::IntoEnumIterator;
-
-fn create_trace_all_steps(domain_size: usize, rng: &mut StdRng) -> DecomposedKeccakTrace {
-    let mut trace = <DecomposedKeccakTrace as DecomposableTracer<KeccakEnv<Fp>>>::new(
-        domain_size,
-        &mut KeccakEnv::<Fp>::default(),
-    );
-    {
-        // 1 block preimages for Sponge(Absorb(Only)), Round(0), and Sponge(Squeeze)
-        for _ in 0..domain_size {
-            // random 1-block preimages
-            let bytelength = rng.gen_range(0..RATE_IN_BYTES);
-            let preimage: Vec<u8> = (0..bytelength).map(|_| rng.gen()).collect();
-            // Initialize the environment and run the interpreter
-            let mut keccak_env = KeccakEnv::<Fp>::new(0, &preimage);
-            while keccak_env.step.is_some() {
-                let step = keccak_env.step.unwrap();
-                // Create the relation witness columns
-                keccak_env.step();
-                match step {
-                    Sponge(Absorb(Only)) | Round(0) | Sponge(Squeeze) => {
-                        // Add the witness row to the circuit
-                        trace.push_row(step, &keccak_env.witness_env.witness.cols);
-                    }
-                    _ => {}
-                }
-            }
-        }
-        // Check there is no need for padding because we reached domain_size rows for these selectors
-        assert!(trace.is_full(Sponge(Absorb(Only))));
-        assert!(trace.is_full(Round(0)));
-        assert!(trace.is_full(Sponge(Squeeze)));
-
-        // Add the columns of the selectors to the circuit
-        trace.set_selector_column::<N_ZKVM_KECCAK_REL_COLS>(Sponge(Absorb(Only)), domain_size);
-        trace.set_selector_column::<N_ZKVM_KECCAK_REL_COLS>(Round(0), domain_size);
-        trace.set_selector_column::<N_ZKVM_KECCAK_REL_COLS>(Sponge(Squeeze), domain_size);
-    }
-    {
-        // 3 block preimages for Sponge(Absorb(First)), Sponge(Absorb(Middle)), and Sponge(Absorb(Last))
-        for _ in 0..domain_size {
-            // random 3-block preimages
-            let bytelength = rng.gen_range(2 * RATE_IN_BYTES..3 * RATE_IN_BYTES);
-            let preimage: Vec<u8> = (0..bytelength).map(|_| rng.gen()).collect();
-            // Initialize the environment and run the interpreter
-            let mut keccak_env = KeccakEnv::<Fp>::new(0, &preimage);
-            while keccak_env.step.is_some() {
-                let step = keccak_env.step.unwrap();
-                // Create the relation witness columns
-                keccak_env.step();
-                match step {
-                    Sponge(Absorb(First)) | Sponge(Absorb(Middle)) | Sponge(Absorb(Last)) => {
-                        // Add the witness row to the circuit
-                        trace.push_row(step, &keccak_env.witness_env.witness.cols);
-                    }
-                    _ => {}
-                }
-            }
-        }
-        // Check there is no need for padding because we reached domain_size rows for these selectors
-        assert!(trace.is_full(Sponge(Absorb(First))));
-        assert!(trace.is_full(Sponge(Absorb(Middle))));
-        assert!(trace.is_full(Sponge(Absorb(Last))));
-
-        // Add the columns of the selectors to the circuit
-        trace.set_selector_column::<N_ZKVM_KECCAK_REL_COLS>(Sponge(Absorb(First)), domain_size);
-        trace.set_selector_column::<N_ZKVM_KECCAK_REL_COLS>(Sponge(Absorb(Middle)), domain_size);
-        trace.set_selector_column::<N_ZKVM_KECCAK_REL_COLS>(Sponge(Absorb(Last)), domain_size);
-        trace
-    }
-}
+use std::collections::HashMap;
 
 #[test]
 fn test_pad_blocks() {
-    let blocks_1 = crate::keccak::pad_blocks::<Fp>(1);
+    let blocks_1 = crate::interpreters::keccak::pad_blocks::<Fp>(1);
     assert_eq!(blocks_1[0], Fp::from(0x00));
     assert_eq!(blocks_1[1], Fp::from(0x00));
     assert_eq!(blocks_1[2], Fp::from(0x00));
     assert_eq!(blocks_1[3], Fp::from(0x00));
     assert_eq!(blocks_1[4], Fp::from(0x81));
 
-    let blocks_136 = crate::keccak::pad_blocks::<Fp>(136);
+    let blocks_136 = crate::interpreters::keccak::pad_blocks::<Fp>(136);
     assert_eq!(blocks_136[0], Fp::from(0x010000000000000000000000u128));
     assert_eq!(blocks_136[1], Fp::from(0x00));
     assert_eq!(blocks_136[2], Fp::from(0x00));
@@ -574,329 +484,4 @@ fn test_keccak_multiplicities() {
             .unwrap()[0],
         2
     );
-}
-
-// Prover/Verifier test includidng the Keccak constraints
-#[test]
-fn test_keccak_prover_constraints() {
-    // guaranteed to have at least 30MB of stack
-    stacker::grow(30 * 1024 * 1024, || {
-        let mut rng = o1_utils::tests::make_test_rng(None);
-        let domain_size = 1 << 8;
-
-        // Generate 3 blocks of preimage data
-        let bytelength = rng.gen_range(2 * RATE_IN_BYTES..RATE_IN_BYTES * 3);
-        let preimage: Vec<u8> = (0..bytelength).map(|_| rng.gen()).collect();
-
-        // Initialize the environment and run the interpreter
-        let mut keccak_env = KeccakEnv::<Fp>::new(0, &preimage);
-
-        // Keep track of the constraints and lookups of the sub-circuits
-        let mut keccak_circuit = <DecomposedKeccakTrace as DecomposableTracer<KeccakEnv<Fp>>>::new(
-            domain_size,
-            &mut keccak_env,
-        );
-
-        while keccak_env.step.is_some() {
-            let step = keccak_env.selector();
-
-            // Run the interpreter, which sets the witness columns
-            keccak_env.step();
-
-            // Add the witness row to the circuit
-            keccak_circuit.push_row(step, &keccak_env.witness_env.witness.cols);
-        }
-        keccak_circuit.pad_witnesses();
-
-        for step in Steps::iter().flat_map(|x| x.into_iter()) {
-            if keccak_circuit.in_circuit(step) {
-                test_completeness_generic_no_lookups::<
-                    N_ZKVM_KECCAK_COLS,
-                    N_ZKVM_KECCAK_REL_COLS,
-                    N_ZKVM_KECCAK_SEL_COLS,
-                    0,
-                    _,
-                >(
-                    keccak_circuit[step].constraints.clone(),
-                    Box::new([]),
-                    keccak_circuit[step].witness.clone(),
-                    domain_size,
-                    &mut rng,
-                );
-            }
-        }
-    });
-}
-
-fn dummy_constraints() -> BTreeMap<Steps, Vec<FoldingCompatibleExpr<KeccakConfig>>> {
-    Steps::iter()
-        .flat_map(|x| x.into_iter())
-        .map(|step| {
-            (
-                step,
-                vec![FoldingCompatibleExpr::<KeccakConfig>::Atom(
-                    FoldingCompatibleExprInner::Constant(Fp::zero()),
-                )],
-            )
-        })
-        .collect()
-}
-
-// (Instance, Witness)
-type KeccakFoldingSide = (
-    <KeccakConfig as FoldingConfig>::Instance,
-    <KeccakConfig as FoldingConfig>::Witness,
-);
-
-// (Step, Left, Right)
-type KeccakFoldingPair = (Steps, KeccakFoldingSide, KeccakFoldingSide);
-
-#[test]
-fn heavy_test_keccak_folding() {
-    use crate::{keccak::folding::KeccakConfig, trace::Foldable, Curve};
-    use ark_poly::{EvaluationDomain, Radix2EvaluationDomain as D};
-    use folding::{checker::Checker, expressions::FoldingCompatibleExpr};
-    use kimchi::curve::KimchiCurve;
-    use mina_poseidon::FqSponge;
-    use poly_commitment::srs::SRS;
-
-    // guaranteed to have at least 30MB of stack
-    stacker::grow(30 * 1024 * 1024, || {
-        let mut rng = o1_utils::tests::make_test_rng(None);
-        let domain_size = 1 << 6;
-
-        let domain = D::<Fp>::new(domain_size).unwrap();
-        let mut srs = SRS::<Curve>::create(domain_size);
-        srs.add_lagrange_basis(domain);
-
-        // Create sponge
-        let mut fq_sponge = BaseSponge::new(Curve::other_curve_sponge_params());
-
-        // Create two instances for each selector to be folded
-        let keccak_trace: [DecomposedKeccakTrace; 2] =
-            array::from_fn(|_| create_trace_all_steps(domain_size, &mut rng));
-        let trace = keccak_trace[0].clone();
-
-        // Store all constraints indexed by Step
-        let constraints = <DecomposedKeccakTrace as Foldable<
-            N_ZKVM_KECCAK_COLS,
-            KeccakConfig,
-            BaseSponge,
-        >>::folding_constraints(&trace);
-
-        // DEFINITIONS OF FUNCTIONS FOR TESTING PURPOSES
-
-        let check_instance_satisfy_constraints =
-            |constraints: &[FoldingCompatibleExpr<KeccakConfig>], side: &KeccakFoldingSide| {
-                let (instance, witness) = side;
-                let checker = Provider::new(instance.clone(), witness.clone());
-                constraints.iter().for_each(|c| {
-                    checker.check(c, domain);
-                });
-            };
-
-        let check_folding =
-            |left: &KeccakFoldingSide,
-             right: &KeccakFoldingSide,
-             constraints: &[FoldingCompatibleExpr<KeccakConfig>],
-             fq_sponge: &mut DefaultFqSponge<Parameters, PlonkSpongeConstantsKimchi>| {
-                // Create the folding scheme ignoring selectors
-                let (scheme, final_constraint) =
-                    FoldingScheme::<KeccakConfig>::new(constraints.to_vec(), &srs, domain, &());
-
-                // Fold both sides and check the constraints ignoring the selector columns
-                let fout =
-                    scheme.fold_instance_witness_pair(left.clone(), right.clone(), fq_sponge);
-
-                // We should always have 0 as the degree of the constraints,
-                // without selectors, they are never higher than 2 in Keccak.
-                assert_eq!(scheme.get_number_of_additional_columns(), 0);
-
-                let checker = ExtendedProvider::new(fout.folded_instance, fout.folded_witness);
-                checker.check(&final_constraint, domain);
-            };
-
-        let check_decomposable_folding_pair =
-            |step: Option<Steps>,
-             left: &KeccakFoldingSide,
-             right: &KeccakFoldingSide,
-             scheme: &DecomposableFoldingScheme<KeccakConfig>,
-             final_constraint: &FoldingCompatibleExpr<KeccakConfig>,
-             quadri_cols: Option<usize>,
-             fq_sponge: &mut DefaultFqSponge<Parameters, PlonkSpongeConstantsKimchi>| {
-                let fout =
-                    scheme.fold_instance_witness_pair(left.clone(), right.clone(), step, fq_sponge);
-
-                let extra_cols = scheme.get_number_of_additional_columns();
-                if let Some(quadri_cols) = quadri_cols {
-                    assert!(extra_cols == quadri_cols);
-                }
-
-                // Check the constraints on the folded circuit applying selectors
-                let checker = ExtendedProvider::<KeccakConfig>::new(
-                    fout.folded_instance,
-                    fout.folded_witness,
-                );
-                checker.check(final_constraint, domain);
-            };
-
-        let check_decomposable_folding =
-            |pair: &KeccakFoldingPair,
-             constraints: BTreeMap<Steps, Vec<FoldingCompatibleExpr<KeccakConfig>>>,
-             common_constraints: Vec<FoldingCompatibleExpr<KeccakConfig>>,
-             quadri_cols: Option<usize>,
-             fq_sponge: &mut DefaultFqSponge<Parameters, PlonkSpongeConstantsKimchi>| {
-                let (step, left, right) = pair;
-                let (dec_scheme, dec_final_constraint) =
-                    DecomposableFoldingScheme::<KeccakConfig>::new(
-                        constraints,
-                        common_constraints,
-                        &srs,
-                        domain,
-                        &(),
-                    );
-                // Subcase A: Check the folded circuit of decomposable folding ignoring selectors (None)
-                check_decomposable_folding_pair(
-                    None,
-                    left,
-                    right,
-                    &dec_scheme,
-                    &dec_final_constraint,
-                    quadri_cols,
-                    fq_sponge,
-                );
-                // Subcase B: Check the folded circuit of decomposable folding applying selectors (Some)
-                check_decomposable_folding_pair(
-                    Some(*step),
-                    left,
-                    right,
-                    &dec_scheme,
-                    &dec_final_constraint,
-                    quadri_cols,
-                    fq_sponge,
-                );
-            };
-
-        let check_decomposable_folding_mix =
-            |steps: (Steps, Steps),
-             fq_sponge: &mut DefaultFqSponge<Parameters, PlonkSpongeConstantsKimchi>| {
-                let (dec_scheme, dec_final_constraint) =
-                    DecomposableFoldingScheme::<KeccakConfig>::new(
-                        constraints.clone(),
-                        vec![],
-                        &srs,
-                        domain,
-                        &(),
-                    );
-                let left = {
-                    let fout = dec_scheme.fold_instance_witness_pair(
-                        keccak_trace[0].to_folding_pair(steps.0, fq_sponge, domain, &srs),
-                        keccak_trace[1].to_folding_pair(steps.0, fq_sponge, domain, &srs),
-                        Some(steps.0),
-                        fq_sponge,
-                    );
-                    let checker = ExtendedProvider::<KeccakConfig>::new(
-                        fout.folded_instance,
-                        fout.folded_witness,
-                    );
-                    (checker.instance, checker.witness)
-                };
-                let right = {
-                    let fout = dec_scheme.fold_instance_witness_pair(
-                        keccak_trace[0].to_folding_pair(steps.1, fq_sponge, domain, &srs),
-                        keccak_trace[1].to_folding_pair(steps.1, fq_sponge, domain, &srs),
-                        Some(steps.1),
-                        fq_sponge,
-                    );
-                    let checker = ExtendedProvider::<KeccakConfig>::new(
-                        fout.folded_instance,
-                        fout.folded_witness,
-                    );
-                    (checker.instance, checker.witness)
-                };
-                let fout = dec_scheme.fold_instance_witness_pair(left, right, None, fq_sponge);
-                let checker = ExtendedProvider::new(fout.folded_instance, fout.folded_witness);
-                checker.check(&dec_final_constraint, domain);
-            };
-
-        // HERE STARTS THE TESTING
-
-        // Sanity checks that the number of constraints are as expected for each step
-        assert_eq!(constraints[&Sponge(Absorb(First))].len(), 332);
-        assert_eq!(constraints[&Sponge(Absorb(Middle))].len(), 232);
-        assert_eq!(constraints[&Sponge(Absorb(Last))].len(), 374);
-        assert_eq!(constraints[&Sponge(Absorb(Only))].len(), 474);
-        assert_eq!(constraints[&Sponge(Squeeze)].len(), 16);
-        assert_eq!(constraints[&Round(0)].len(), 389);
-
-        // Total number of Keccak constraints of degree higher than 2 (should be 0)
-        let total_deg_higher_2 = Steps::iter()
-            .flat_map(|x| x.into_iter())
-            .fold(0, |acc, step| {
-                acc + trace[step]
-                    .constraints
-                    .iter()
-                    .filter(|c| c.degree(1, 0) > 2)
-                    .count()
-            });
-        assert_eq!(total_deg_higher_2, 0);
-
-        // Check folding constraints of individual steps ignoring selectors
-        for step in Steps::iter().flat_map(|x| x.into_iter()) {
-            // BTreeMap with constraints of this step
-            let mut step_constraints = BTreeMap::new();
-            step_constraints.insert(step, constraints[&step].clone());
-
-            // Create sides for folding
-            let left = keccak_trace[0].to_folding_pair(step, &mut fq_sponge, domain, &srs);
-            let right = keccak_trace[1].to_folding_pair(step, &mut fq_sponge, domain, &srs);
-
-            // CASE 0: Check instances satisfy the constraints, without folding them
-            check_instance_satisfy_constraints(&constraints[&step], &left);
-            check_instance_satisfy_constraints(&constraints[&step], &right);
-
-            // CASE 1: Check constraints on folded circuit ignoring selectors with `FoldingScheme`
-            check_folding(&left, &right, &constraints[&step], &mut fq_sponge);
-
-            // CASE 2: Check that `DecomposableFoldingScheme` works when passing the dummy zero constraint
-            //         to each step, and an empty list of common constraints.
-            let pair = (step, left, right);
-            check_decomposable_folding(&pair, dummy_constraints(), vec![], Some(0), &mut fq_sponge);
-
-            // CASE 3: Using a separate `DecomposableFoldingScheme` for each step, check each step
-            //         constraints using a dummy BTreeMap of `vec[0]` per-step constraints and
-            //         common constraints set to each selector's constraints.
-            check_decomposable_folding(
-                &pair,
-                dummy_constraints(),
-                step_constraints[&step].clone(),
-                Some(0),
-                &mut fq_sponge,
-            );
-
-            // CASE 4: Using the same `DecomposableFoldingScheme` for all steps, initialized with a real
-            //         BTreeMap of only the current step, and common constraints set to `vec[]`, check
-            //         the folded circuit
-            check_decomposable_folding(
-                &pair,
-                step_constraints.clone(),
-                vec![],
-                None,
-                &mut fq_sponge,
-            );
-
-            // CASE 5: Using the same `DecomposableFoldingScheme` for all steps, initialized with a real
-            //         BTreeMap of constraints per-step, and common constraints set to `vec[]`, check
-            //         the folded circuit
-            check_decomposable_folding(
-                &pair,
-                constraints.clone(),
-                vec![],
-                Some(151),
-                &mut fq_sponge,
-            );
-        }
-        // CASE 6: Fold mixed steps together and check the final constraints
-        check_decomposable_folding_mix((Sponge(Absorb(First)), Round(0)), &mut fq_sponge);
-    });
 }
