+++ o1vm/src/interpreters/mips/witness.rs
@@ -3,20 +3,23 @@ use crate::{
         Hint, Meta, Page, Start, State, StepFrequency, VmConfiguration, PAGE_ADDRESS_MASK,
         PAGE_ADDRESS_SIZE, PAGE_SIZE,
     },
-    keccak::environment::KeccakEnv,
-    lookups::Lookup,
-    mips::{
-        column::{
-            ColumnAlias as Column, MIPS_BYTE_COUNTER_OFF, MIPS_CHUNK_BYTES_LEN,
-            MIPS_END_OF_PREIMAGE_OFF, MIPS_HASH_COUNTER_OFF, MIPS_HAS_N_BYTES_OFF,
-            MIPS_LENGTH_BYTES_OFF, MIPS_NUM_BYTES_READ_OFF, MIPS_PREIMAGE_BYTES_OFF,
-            MIPS_PREIMAGE_CHUNK_OFF, MIPS_PREIMAGE_KEY,
+    interpreters::{
+        keccak::environment::KeccakEnv,
+        mips::{
+            column::{
+                ColumnAlias as Column, MIPS_BYTE_COUNTER_OFF, MIPS_CHUNK_BYTES_LEN,
+                MIPS_END_OF_PREIMAGE_OFF, MIPS_HASH_COUNTER_OFF, MIPS_HAS_N_BYTES_OFF,
+                MIPS_LENGTH_BYTES_OFF, MIPS_NUM_BYTES_READ_OFF, MIPS_PREIMAGE_BYTES_OFF,
+                MIPS_PREIMAGE_CHUNK_OFF, MIPS_PREIMAGE_KEY,
+            },
+            interpreter::{
+                self, ITypeInstruction, Instruction, InterpreterEnv, JTypeInstruction,
+                RTypeInstruction,
+            },
+            registers::Registers,
         },
-        interpreter::{
-            self, ITypeInstruction, Instruction, InterpreterEnv, JTypeInstruction, RTypeInstruction,
-        },
-        registers::Registers,
     },
+    lookups::Lookup,
     preimage_oracle::PreImageOracleT,
 };
 use ark_ff::Field;
