+++ o1vm/src/legacy/trace.rs
@@ -6,7 +6,7 @@
 //! selected using "selectors".
 
 use crate::{
-    folding::{BaseField, FoldingInstance, FoldingWitness, ScalarField},
+    legacy::folding::{BaseField, FoldingInstance, FoldingWitness, ScalarField},
     lookups::Lookup,
     E,
 };
@@ -319,6 +319,246 @@ where
             0
         } else {
             self.trace.get_mut(&selector).unwrap().pad_dummy(())
+        }
+    }
+}
+
+pub mod keccak {
+    use std::{array, collections::BTreeMap};
+
+    use ark_ff::Zero;
+    use kimchi_msm::witness::Witness;
+    use strum::IntoEnumIterator;
+
+    use crate::{
+        interpreters::keccak::{
+            column::{Steps, N_ZKVM_KECCAK_COLS, N_ZKVM_KECCAK_REL_COLS},
+            environment::KeccakEnv,
+            standardize,
+        },
+        legacy::{
+            folding::{keccak::KeccakConfig, ScalarField},
+            trace::{DecomposableTracer, DecomposedTrace, Trace, Tracer},
+        },
+    };
+
+    /// A Keccak instruction trace
+    pub type KeccakTrace = Trace<N_ZKVM_KECCAK_COLS, KeccakConfig>;
+    /// The Keccak circuit trace
+    pub type DecomposedKeccakTrace = DecomposedTrace<N_ZKVM_KECCAK_COLS, KeccakConfig>;
+
+    impl DecomposableTracer<KeccakEnv<ScalarField<KeccakConfig>>> for DecomposedKeccakTrace {
+        fn new(domain_size: usize, env: &mut KeccakEnv<ScalarField<KeccakConfig>>) -> Self {
+            let mut circuit = Self {
+                domain_size,
+                trace: BTreeMap::new(),
+            };
+            for step in Steps::iter().flat_map(|step| step.into_iter()) {
+                circuit
+                    .trace
+                    .insert(step, KeccakTrace::init(domain_size, step, env));
+            }
+            circuit
+        }
+
+        fn pad_witnesses(&mut self) {
+            for opcode in Steps::iter().flat_map(|opcode| opcode.into_iter()) {
+                if self.in_circuit(opcode) {
+                    self.trace.get_mut(&opcode).unwrap().pad_dummy(());
+                }
+            }
+        }
+    }
+
+    impl Tracer<N_ZKVM_KECCAK_REL_COLS, KeccakConfig, KeccakEnv<ScalarField<KeccakConfig>>>
+        for KeccakTrace
+    {
+        type Selector = ();
+
+        fn init(
+            domain_size: usize,
+            selector: Steps,
+            _env: &mut KeccakEnv<ScalarField<KeccakConfig>>,
+        ) -> Self {
+            // Make sure we are using the same round number to refer to round steps
+            let step = standardize(selector);
+            Self {
+                domain_size,
+                witness: Witness {
+                    cols: Box::new(std::array::from_fn(|_| Vec::with_capacity(domain_size))),
+                },
+                constraints: KeccakEnv::constraints_of(step),
+                lookups: KeccakEnv::lookups_of(step),
+            }
+        }
+
+        fn push_row(
+            &mut self,
+            _selector: Self::Selector,
+            row: &[ScalarField<KeccakConfig>; N_ZKVM_KECCAK_REL_COLS],
+        ) {
+            for (i, value) in row.iter().enumerate() {
+                if self.witness.cols[i].len() < self.witness.cols[i].capacity() {
+                    self.witness.cols[i].push(*value);
+                }
+            }
+        }
+
+        fn pad_with_row(
+            &mut self,
+            _selector: Self::Selector,
+            row: &[ScalarField<KeccakConfig>; N_ZKVM_KECCAK_REL_COLS],
+        ) -> usize {
+            let len = self.witness.cols[0].len();
+            assert!(len <= self.domain_size);
+            let rows_to_add = self.domain_size - len;
+            // When we reach the domain size, we don't need to pad anymore.
+            for _ in 0..rows_to_add {
+                self.push_row((), row);
+            }
+            rows_to_add
+        }
+
+        fn pad_with_zeros(&mut self, _selector: Self::Selector) -> usize {
+            let len = self.witness.cols[0].len();
+            assert!(len <= self.domain_size);
+            let rows_to_add = self.domain_size - len;
+            // When we reach the domain size, we don't need to pad anymore.
+            for col in self.witness.cols.iter_mut() {
+                col.extend((0..rows_to_add).map(|_| ScalarField::<KeccakConfig>::zero()));
+            }
+            rows_to_add
+        }
+
+        fn pad_dummy(&mut self, _selector: Self::Selector) -> usize {
+            // We keep track of the first row of the non-empty witness, which is a real step witness.
+            let row = array::from_fn(|i| self.witness.cols[i][0]);
+            self.pad_with_row(_selector, &row)
+        }
+    }
+}
+
+pub mod mips {
+    use crate::{
+        interpreters::mips::{
+            column::{N_MIPS_COLS, N_MIPS_REL_COLS},
+            constraints::Env,
+            interpreter::{interpret_instruction, Instruction},
+        },
+        legacy::{
+            folding::{mips::DecomposableMIPSFoldingConfig, ScalarField},
+            trace::{DecomposableTracer, DecomposedTrace, Trace, Tracer},
+        },
+    };
+    use ark_ff::Zero;
+    use kimchi_msm::witness::Witness;
+    use std::{array, collections::BTreeMap};
+    use strum::IntoEnumIterator;
+
+    /// The MIPS instruction trace
+    pub type MIPSTrace = Trace<N_MIPS_COLS, DecomposableMIPSFoldingConfig>;
+    /// The MIPS circuit trace
+    pub type DecomposedMIPSTrace = DecomposedTrace<N_MIPS_COLS, DecomposableMIPSFoldingConfig>;
+
+    impl DecomposableTracer<Env<ScalarField<DecomposableMIPSFoldingConfig>>> for DecomposedMIPSTrace {
+        fn new(
+            domain_size: usize,
+            env: &mut Env<ScalarField<DecomposableMIPSFoldingConfig>>,
+        ) -> Self {
+            let mut circuit = Self {
+                domain_size,
+                trace: BTreeMap::new(),
+            };
+            for instr in Instruction::iter().flat_map(|step| step.into_iter()) {
+                circuit
+                    .trace
+                    .insert(instr, <MIPSTrace>::init(domain_size, instr, env));
+            }
+            circuit
+        }
+
+        fn pad_witnesses(&mut self) {
+            for opcode in Instruction::iter().flat_map(|opcode| opcode.into_iter()) {
+                self.trace.get_mut(&opcode).unwrap().pad_dummy(());
+            }
+        }
+    }
+
+    impl
+        Tracer<
+            N_MIPS_REL_COLS,
+            DecomposableMIPSFoldingConfig,
+            Env<ScalarField<DecomposableMIPSFoldingConfig>>,
+        > for MIPSTrace
+    {
+        type Selector = ();
+
+        fn init(
+            domain_size: usize,
+            instr: Instruction,
+            env: &mut Env<ScalarField<DecomposableMIPSFoldingConfig>>,
+        ) -> Self {
+            interpret_instruction(env, instr);
+
+            let trace = Self {
+                domain_size,
+                witness: Witness {
+                    cols: Box::new(std::array::from_fn(|_| Vec::with_capacity(domain_size))),
+                },
+                constraints: env.constraints.clone(),
+                lookups: env.lookups.clone(),
+            };
+            env.scratch_state_idx = 0; // Reset the scratch state index for the next instruction
+            env.constraints = vec![]; // Clear the constraints for the next instruction
+            env.lookups = vec![]; // Clear the lookups for the next instruction
+
+            trace
+        }
+
+        fn push_row(
+            &mut self,
+            _selector: Self::Selector,
+            row: &[ScalarField<DecomposableMIPSFoldingConfig>; N_MIPS_REL_COLS],
+        ) {
+            for (i, value) in row.iter().enumerate() {
+                if self.witness.cols[i].len() < self.witness.cols[i].capacity() {
+                    self.witness.cols[i].push(*value);
+                }
+            }
+        }
+
+        fn pad_with_row(
+            &mut self,
+            _selector: Self::Selector,
+            row: &[ScalarField<DecomposableMIPSFoldingConfig>; N_MIPS_REL_COLS],
+        ) -> usize {
+            let len = self.witness.cols[0].len();
+            assert!(len <= self.domain_size);
+            let rows_to_add = self.domain_size - len;
+            // When we reach the domain size, we don't need to pad anymore.
+            for _ in 0..rows_to_add {
+                self.push_row(_selector, row);
+            }
+            rows_to_add
+        }
+
+        fn pad_with_zeros(&mut self, _selector: Self::Selector) -> usize {
+            let len = self.witness.cols[0].len();
+            assert!(len <= self.domain_size);
+            let rows_to_add = self.domain_size - len;
+            // When we reach the domain size, we don't need to pad anymore.
+            for col in self.witness.cols.iter_mut() {
+                col.extend(
+                    (0..rows_to_add).map(|_| ScalarField::<DecomposableMIPSFoldingConfig>::zero()),
+                );
+            }
+            rows_to_add
+        }
+
+        fn pad_dummy(&mut self, _selector: Self::Selector) -> usize {
+            // We keep track of the first row of the non-empty witness, which is a real step witness.
+            let row = array::from_fn(|i| self.witness.cols[i][0]);
+            self.pad_with_row(_selector, &row)
         }
     }
 }
