+++ o1vm/src/lib.rs
@@ -1,38 +1,24 @@
-/// Domain size shared by the Keccak evaluations, MIPS evaluation and main
-/// program.
-pub const DOMAIN_SIZE: usize = 1 << 15;
-
 /// Modules mimicking the defined structures used by Cannon CLI.
 pub mod cannon;
 
 /// A CLI mimicking the Cannon CLI.
 pub mod cannon_cli;
 
-/// Integration with folding. Contains common trait implementations to be used
-/// by each circuit.
-pub mod folding;
+pub mod interpreters;
 
-/// Implementation of Keccak used by the zkVM.
-pub mod keccak;
+/// Legacy implementation of the recursive proof composition.
+/// It does use the folding and ivc libraries defined in this monorepo, and aims
+/// to be compatible with Ethereum natively, using the curve bn254.
+pub mod legacy;
 
 /// Instantiation of the lookups for the VM project.
 pub mod lookups;
 
-/// MIPS interpreter.
-pub mod mips;
-
 /// Preimage oracle interface used by the zkVM.
 pub mod preimage_oracle;
 
-/// Proof system of the zkVM.
-pub mod proof;
-
 /// The RAM lookup argument.
 pub mod ramlookup;
-
-/// Abstract execution traces, possible long, that can be folded.
-/// A trace is a sequence of data points organized in a 2D array, constrained.
-pub mod trace;
 
 use ark_ec::bn::Bn;
 use kimchi::circuits::expr::{ConstantExpr, Expr};
@@ -60,6 +46,7 @@ pub use ramlookup::{LookupMode as RAMLookupMode, RAMLookup};
 pub(crate) type E<F> = Expr<ConstantExpr<F>, Column>;
 
 // Instantiate it with the desired group and field
+// FIXME: move this into legacy::main.rs
 
 /// Scalar field of BN254
 pub type Fp = ark_bn254::Fr;
